Extract server actions into their own files
